feat: rewrite indexes to TS
