+++ native-stack/package.json
@@ -2,5 +2,5 @@
   "main": "../lib/commonjs/native-stack/index",
   "module": "../lib/module/native-stack/index",
   "react-native": "../src/native-stack/index",
-  "types": "../lib/typescript/index.d.ts"
+  "types": "../lib/typescript/native-stack/index.d.ts"
 }
